+++ inst/sql/impact_method2_metadata/burden_outcomes.sql
@@ -1,0 +1,2 @@
+SELECT id, code
+FROM burden_outcome